First Commit: Created a products app api integrating with mongoDb
The app includes a product model, a controller with functions such as getAllProducts, updateProductInfo, and deleteProduct, as well as a router with routes that utilize the controller functions. Additionally, it connects to a database using a provided URL.

The required packages, dotenv, express, mongoose, and nodemon, have been added to the project.